+++ include/core/managers/BaseManager.hpp
@@ -11,5 +11,10 @@ public:
     BaseManager();
     virtual ~BaseManager();
 
-    virtual void Bind(std::shared_ptr<GameObject> gameObject) = 0;
+    virtual void Bind(std::shared_ptr<GameObject> gameObject);
+    
+    // Common methods with default implementations
+    virtual void Init();
+    virtual void Update(float deltaTime);
+    virtual bool IsEmpty();
 };