+++ include/core/managers/EnemyManager.hpp
@@ -9,9 +9,5 @@ public:
     EnemyManager();
     ~EnemyManager();
 
-    void Init();
-    virtual void Bind(std::shared_ptr<GameObject> gameObject) override;
     void SpawnEnemy(const FVector2& position, const FVector2& size, const Color color, float accelerationIndex);
-    void Update(float deltaTime);
-    bool IsEmpty();
 };